reverted patrol tests
