+++ test/chat_bubble_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:geolocator/geolocator.dart';
@@ -20,19 +22,14 @@ import 'package:xcnav/providers/group.dart';
 import 'package:xcnav/providers/my_telemetry.dart';
 import 'package:xcnav/providers/plans.dart';
 import 'package:xcnav/providers/profile.dart';
+import 'package:xcnav/settings_service.dart';
 import 'package:xcnav/providers/weather.dart';
 import 'package:xcnav/providers/wind.dart';
-import 'package:xcnav/settings_service.dart';
 
 import 'mock_providers.dart';
 
 void main() {
-  SharedPreferences.setMockInitialValues({});
-  SharedPreferences.getInstance().then((prefs) {
-    settingsMgr = SettingsMgr(prefs);
-  });
-
-  Widget makeApp() {
+  Widget makeApp(ActivePlan activePlan, MockPlans plans, Completer<MockClient> client) {
     return MultiProvider(providers: [
       ChangeNotifierProvider(
         create: (_) => MyTelemetry(),
@@ -43,15 +40,16 @@ void main() {
         lazy: false,
       ),
       ChangeNotifierProvider(
-        create: (_) => Wind(),
+        create: (context) => Wind(),
         lazy: false,
       ),
       ChangeNotifierProvider(
-        create: (_) => ActivePlan(),
+        create: (_) => activePlan,
         lazy: false,
       ),
       ChangeNotifierProvider(
-        create: (_) => Plans(),
+        // ignore: unnecessary_cast
+        create: (_) => plans as Plans,
         lazy: false,
       ),
       ChangeNotifierProvider(
@@ -71,8 +69,12 @@ void main() {
         lazy: false,
       ),
       ChangeNotifierProvider(
-        // ignore: unnecessary_cast
-        create: (context) => MockClient(context) as Client,
+        create: (context) {
+          final fakeClient = MockClient(context);
+          client.complete(fakeClient);
+          // ignore: unnecessary_cast
+          return fakeClient as Client;
+        },
         lazy: false,
       )
     ], child: const XCNav());
@@ -83,9 +85,16 @@ void main() {
     final fontLoader = FontLoader('roboto-condensed')..addFont(flamante);
     await fontLoader.load();
   });
+
   patrolTest(
-    'Flight Logs: loads with no logs',
+    'Check chat bubble appears and clears',
     ($) async {
+      SharedPreferences.setMockInitialValues({});
+      SharedPreferences.getInstance().then((prefs) {
+        settingsMgr = SettingsMgr(prefs);
+        settingsMgr.showAirspaceOverlay.value = false;
+      });
+
       // --- Setup stubs and initial configs
       GeolocatorPlatform.instance = MockGeolocatorPlatform();
       perm_handler_plat.PermissionHandlerPlatform.instance = MockPermissionHandlerPlatform();
@@ -101,22 +110,69 @@ void main() {
         "profile.id": "1234",
         "profile.secretID": "1234abcd",
       });
+      final activePlan = ActivePlan();
+      final plans = MockPlans();
+      final clientCompleter = Completer<MockClient>();
 
       // --- Build App
-      await $.pumpWidget(makeApp());
+      await $.pumpWidget(makeApp(activePlan, plans, clientCompleter));
       await $.waitUntilExists($(Scaffold));
 
-      // --- Open flight logs screen
-      await $.tester.tapAt($.tester.getBottomLeft($(MaterialApp)) + const Offset(30, -30));
-      await $.pump(const Duration(seconds: 2));
-      await $.tester.drag($("ADSB-in"), const Offset(0, -300));
-      await $.pump(const Duration(seconds: 2));
-      await $("Log").tap(settlePolicy: SettlePolicy.noSettle);
+      final client = await clientCompleter.future;
 
-      // // --- Select Stats tile
-      await $.waitUntilExists($("Entries"));
-      await $("Stats").tap(settlePolicy: SettlePolicy.noSettle);
-      await $.pump(const Duration(seconds: 2));
+      // --- Join a group
+      client.handleAuthResponse({
+        "status": 0,
+        "secretToken": "1234abcd",
+        "pilot_id": "1234",
+        "pilotMetaHash": "cb3b4",
+        "apiVersion": 7,
+        "group_id": "6f3a49"
+      });
+
+      client.handleGroupInfoResponse({
+        "status": 0,
+        "group_id": "6f3a49",
+        "pilots": [
+          {"id": "otherPilotID1234", "name": "bender", "avatarHash": "b96e3cff738b67359e2896db92a11284"}
+        ],
+        "waypoints": {
+          "ld6k9yb82t7h61": {
+            "id": "ld6k9yb82t7h61",
+            "name": "Hospital Canyon",
+            "latlng": [
+              [37.55983, -121.37429]
+            ],
+            "icon": null,
+            "color": 4278190080
+          },
+        }
+      });
+
+      // --- Test a short message
+      client.handleChatMessage({
+        "timestamp": DateTime.now().millisecondsSinceEpoch,
+        "group_id": "6f3a49",
+        "pilot_id": "otherPilotID1234",
+        "text": "testing",
+        "emergency": false
+      });
+
+      await $.waitUntilVisible($("testing"));
+
+      // --- Test a long message
+      client.handleChatMessage({
+        "timestamp": DateTime.now().millisecondsSinceEpoch,
+        "group_id": "6f3a49",
+        "pilot_id": "otherPilotID1234",
+        "text": "ok, this is a long message example. It should line-wrap in the display.",
+        "emergency": false
+      });
+
+      await $
+          .waitUntilVisible($("ok, this is a long message example. It should line-wrap in the display."))
+          .tap(settlePolicy: SettlePolicy.noSettle);
+      await $.pump(const Duration(seconds: 30));
     },
   );
 }
