+++ test/latlng_editor_test.dart
